Add label slot to statement structure
Yes, I know, this should really be on the statement-block structure, but the label is introduced with the statement that introducs the block, not the block body.

Pathological case:
foo:
do.

This way, the above parses as a labeled statement "DO", which the compiler can deal with, rather than having no block to associate the label with.